refactor/move typeorm outside global module
